Allow custom smart attributes via env var
Try to accomplish https://github.com/prometheus-community/node-exporter-textfile-collector-scripts/pull/103
Initial credits to ZigZagT.

Move the env var (SMARTMONATTRS) to the top of the list, otherwise it does not work. (Why?)
The env var should contain the custom attributes separated by a space. The space has to be replaced by a new line. An improvement would be to do this before building the full attributes list.

Signed-off-by: SaturnusDJ <SaturnusDJ@users.noreply.github.com>